Vowel Space Ratio
Based on the tracked formants F1 and F2 for the voiced regions of
speech we compute the vowel space for each recorded subject
individually. We define the vowel space as the frequency region covered
by the triangle in the two dimensional frequency space spanned by F1
and F2 for the vowels /i/ (as in heed), /a/ (as in hod), and /u/ (as in
who’d).

CONFLICT: kmeans of voicebox is not used here and caused a conflict
when running startup.m first. maybe we should remove unused functions
of voicebox to avoid these issues in the future and only include
important ones.